sweep: #7895 Standard naming for TokenManager service classes
