+++ src/DIRAC/FrameworkSystem/ConfigTemplate.cfg
@@ -25,6 +25,20 @@ Services
       storeHostInfo = Operator
     }
   }
+  ##BEGIN TokenManager:
+  # Section to describe TokenManager system
+  TokenManager
+  {
+    Port = 9181
+    # Description of rules for access to methods
+    Authorization
+    {
+      # Settings by default:
+      Default = authenticated
+      getUsersTokensInfo = ProxyManagement
+    }
+  }
+  ##END
   ##BEGIN TornadoTokenManager:
   # Section to describe TokenManager system
   TornadoTokenManager
